requirements.txt
openai
numpy
scikit-learn
umap-learn
matplotlib
tqdm
jupyter
transformers
torch
accelerate
sentence-transformers

